Create a popover used for creating new subjects

+++ apps/admin-panel/src/components/subjects/subjects.tsx
@@ -23,7 +23,7 @@ import Async from "@/better-router/async";
 import { Input } from "../ui/input";
 import { Textarea } from "../ui/textarea";
 import i18n from "@/i18n";
-import { useNavigate } from "react-router";
+import { useNavigate, useRevalidator } from "react-router";
 import {
   AlertDialog,
   AlertDialogAction,
@@ -33,10 +33,12 @@ import {
   AlertDialogTitle,
   AlertDialogTrigger,
 } from "../ui/alert-dialog";
+import { Popover, PopoverContent, PopoverTrigger } from "../ui/popover";
 
 export default function Subjects() {
   const loaderData = useLoader<typeof subjectsLoader>();
   const navigate = useNavigate();
+  const { revalidate } = useRevalidator();
 
   const [currentPage, setCurrentPage] = useState<Promise<
     Schema<"SimpleSubjectResponseDto">[]
@@ -63,8 +65,9 @@ export default function Subjects() {
     const response = sendAPIRequest("/subject", {
       method: "get",
       parameters: {
-        limit: 15,
-        offset: 15 * (pageCount - 1),
+        limit: 10,
+        offset: 10 * (pageCount - 1),
+        languageCode: i18n.language,
       },
     }).then((x) => {
       isLoadingData.current = false;
@@ -82,8 +85,9 @@ export default function Subjects() {
     const response = sendAPIRequest("/subject", {
       method: "get",
       parameters: {
-        limit: 15,
-        offset: 15 * (pageCount + 1),
+        limit: 10,
+        offset: 10 * (pageCount + 1),
+        languageCode: i18n.language,
       },
     }).then((x) => {
       isLoadingData.current = false;
@@ -99,6 +103,9 @@ export default function Subjects() {
     newName: string,
     newDesc: string
   ) {
+    if (isWaitingForResponse.current) return;
+    isWaitingForResponse.current = true;
+
     const response = await sendAPIRequest("/subject/translation", {
       method: "put",
       payload: {
@@ -109,10 +116,15 @@ export default function Subjects() {
       },
     });
 
+    isWaitingForResponse.current = false;
     if (response.code !== "No Content") alert(response);
+    else alert("Uspesno izmenjen predmet");
   }
 
   async function handleDelete(subject: Schema<"SimpleSubjectResponseDto">) {
+    if (isWaitingForResponse.current) return;
+    isWaitingForResponse.current = true;
+
     const response = await sendAPIRequest("/subject/{id}", {
       method: "delete",
       parameters: {
@@ -120,15 +132,43 @@ export default function Subjects() {
       },
     });
 
+    isWaitingForResponse.current = false;
     if (response.code !== "No Content") alert(response);
+    else revalidate();
+  }
+
+  const [isCreationPopoverOpen, setIsCreationPopoverOpen] = useState(false);
+  const isWaitingForResponse = useRef(false);
+  async function handleCreate(name: string, desc: string) {
+    if (isWaitingForResponse.current) return;
+    isWaitingForResponse.current = true;
+
+    const response = await sendAPIRequest("/subject", {
+      method: "post",
+      payload: {
+        name: name,
+        description: desc,
+        languageCode: i18n.language,
+      },
+    });
+
+    isWaitingForResponse.current = false;
+    if (response.code === "Created") {
+      setIsCreationPopoverOpen(false);
+      revalidate();
+    } else alert(response);
   }
 
   return (
-    <Table>
+    <Table className="mt-8">
       <TableHeader>
         <TableRow className="flex">
-          <TableCell className="flex-1/5">Ime</TableCell>
-          <TableCell className="flex-4/5">Opis</TableCell>
+          <TableCell className="flex-1/5 text-3xl font-bold text-start mr-8">
+            Ime
+          </TableCell>
+          <TableCell className="flex-4/5 text-3xl font-bold text-start">
+            Opis
+          </TableCell>
         </TableRow>
       </TableHeader>
 
@@ -138,12 +178,18 @@ export default function Subjects() {
             {(page) =>
               page.map((subject) => (
                 <TableRow key={subject.id} className="flex items-center">
-                  <TableCell className="flex-1/5">
-                    <Input defaultValue={subject.name} />
+                  <TableCell className="flex-1/5 min-w-1/5 mr-8">
+                    <Input
+                      defaultValue={subject.name}
+                      className="text-lg min-h-16"
+                    />
                   </TableCell>
 
-                  <TableCell className="flex-4/5">
-                    <Textarea defaultValue={subject.description} />
+                  <TableCell className="flex-4/5 mr-8">
+                    <Textarea
+                      defaultValue={subject.description}
+                      className="text-lg min-h-16"
+                    />
                   </TableCell>
 
                   <TableCell>
@@ -175,14 +221,17 @@ export default function Subjects() {
                         <Trash2 className="min-w-full min-h-full group-hover:animate-spin group-hover:text-red-600 transition-colors" />
                       </AlertDialogTrigger>
 
-                      <AlertDialogContent>
-                        <AlertDialogTitle>
+                      <AlertDialogContent className="min-w-max gap-12">
+                        <AlertDialogTitle className="text-2xl min-w-max">
                           Da li ste sigurni da zelite da obrisete ovaj predmet?
                         </AlertDialogTitle>
 
-                        <AlertDialogFooter>
-                          <AlertDialogCancel>Otkazi</AlertDialogCancel>
+                        <AlertDialogFooter className="gap-4">
+                          <AlertDialogCancel className="min-w-32 h-16 text-xl">
+                            Otkazi
+                          </AlertDialogCancel>
                           <AlertDialogAction
+                            className="min-w-32 h-16 text-xl"
                             onClick={() => handleDelete(subject)}
                           >
                             Obrisi
@@ -210,17 +259,72 @@ export default function Subjects() {
 
       <TableFooter>
         <TableRow>
-          <TableCell colSpan={3}>
+          <TableCell colSpan={3} className="w-full flex justify-center">
+            <Popover
+              open={isCreationPopoverOpen}
+              onOpenChange={setIsCreationPopoverOpen}
+            >
+              <PopoverTrigger className="min-w-96 min-h-12 border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground">
+                Dodaj predmet
+              </PopoverTrigger>
+
+              <PopoverContent className="flex flex-col gap-4 p-8 min-w-max">
+                <h1>Dodaj predmet</h1>
+                <Input
+                  className="min-w-96 min-h-12 border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground"
+                  placeholder="Ime predmeta"
+                />
+                <Textarea
+                  className="min-w-96 min-h-12 border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground"
+                  placeholder="Opis predmeta"
+                />
+                <Button
+                  className="min-w-96 min-h-16 border border-input bg-background text-muted-foreground shadow-sm hover:bg-accent hover:text-accent-foreground"
+                  onClick={(e) => {
+                    const target = e.target as HTMLElement;
+                    const name = (
+                      target.previousSibling
+                        ?.previousSibling as HTMLInputElement
+                    )?.value?.trim();
+                    const desc = (
+                      target.previousSibling as HTMLInputElement
+                    )?.value?.trim();
+
+                    if (name && desc) handleCreate(name, desc);
+                  }}
+                >
+                  Sacuvaj
+                </Button>
+              </PopoverContent>
+            </Popover>
+          </TableCell>
+        </TableRow>
+
+        <TableRow>
+          <TableCell colSpan={3} className="w-full flex justify-center">
             <div className="flex items-center gap-2">
-              <Button onClick={handlePrevClick}>
-                <ChevronLeft />
+              <Button
+                onClick={handlePrevClick}
+                disabled={pageCount === 0}
+                className="min-w-12 min-h-12 p-2"
+              >
+                <ChevronLeft className="min-w-full min-h-full" />
               </Button>
 
-              <b>Page count: {pageCount}</b>
-
-              <Button onClick={handleNextClick}>
-                <ChevronRight />
-              </Button>
+              <p className="text-xl">Page count: {pageCount + 1}</p>
+              {currentPage && (
+                <Async await={currentPage}>
+                  {(x) => (
+                    <Button
+                      onClick={handleNextClick}
+                      disabled={(x?.length ?? 0) === 0}
+                      className="min-w-12 min-h-12 p-2"
+                    >
+                      <ChevronRight className="min-w-full min-h-full" />
+                    </Button>
+                  )}
+                </Async>
+              )}
             </div>
           </TableCell>
         </TableRow>
